Docker/requirements.txt
Flask==3.0.0
ultralytics
Pillow
gunicorn